requirements.txt
google-generativeai
python-docx
google-api-python-client
firebase-admin
requests
fastapi
uvicorn
slowapi
openai
python-multipart
pydantic
python-dotenv
google-cloud-storage
jinja2
PyPDF2
chromadb
openai